game basic framework
This framework consists of four classes. 
First is the “Window.java”; here is created a window and its properties. Second one is the “Canvas.java”; here is created a panel on which we will draw and listen for keyboard and mouse events. The third is the "Framework.java"; this class controls the game, update and draws it on the screen. The last, fourth class is the "Game.java"; this is the actual game that is created and controlled by "Framework.java".